Setup actor movement loop

See merge request course-work/swen225/2023/project1/team6/larry-crofts-adventures!35

+++ src/main/java/nz/ac/wgtn/swen225/lc/app/gui/GameInfoController.java
@@ -86,7 +86,7 @@ public class GameInfoController {
    * @param game   The game that is relevant to the game bar.
    */
   public void updateUi(Domain domain, Game game) {
-    timeLeft.setText(Integer.toString(game.getTimeLeft()));
+    timeLeft.setText(Long.toString(game.getTimeLeft()));
     treasureLeft.setText(Integer.toString(domain.getPlayer().getTreasuresLeft()));
 
     List<Item> playerInventory = domain.getPlayer().getInventory();
